requirements.txt
anyascii==0.3.1
APScheduler==3.10.0
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
boto3==1.26.76
botocore==1.29.76
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==4.1.7
django-bootstrap-form==3.4
django-crum==0.7.9
django-discover-runner==1.0
django-filter==22.1
django-import-export==3.1.0
django-logentry-admin==1.1.0
django-modelcluster==6.0
django-parsley==0.7
django-permissionedforms==0.1
django-role-permissions==3.1.1
django-session-security==2.6.7
django-simple-history==3.2.0
django-storages==1.13.2
django-taggit==3.1.0
django-treebeard==4.6.1
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fonttools==4.38.0
googlemaps==4.10.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
jmespath==1.0.1
kiwisolver==1.4.4
l18n==2021.3
lxml==4.9.2
MarkupPy==1.14
matplotlib==3.7.0
numpy==1.24.2
odfpy==1.4.1
opencv-python-headless==4.7.0.68
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pluggy==1.0.0
point==0.0.1
psycopg2==2.9.5
py==1.11.0
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
python-decouple==3.7
python3-saml==1.15.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
shape==0.1.1
Shapely @ file:///C:/Users/17372/Desktop/capstone310/Local_env/uno-cpi/Shapely-1.8.2-cp310-cp310-win_amd64.whl
six==1.11.0
soupsieve==2.4
sqlparse==0.4.3
tablib==3.3.0
telepath==0.3
toml==0.10.2
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
wagtail==4.2
webencodings==0.5.1
whitenoise==6.3.0
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.8
xlwt==1.3.0
xmlsec==1.3.13
